pyproject.toml
[project]
name = "databricks-ai-bridge"
version = "0.9.0.dev0"
description = "Official Python library for Databricks AI support"
authors = [
    { name="Databricks", email="agent-feedback@databricks.com" },
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "typing_extensions",
  "pydantic",
  "databricks-sdk>=0.49.0",
  "pandas",
  "tiktoken>=0.8.0",
  "tabulate>=0.9.0",
  "mlflow-skinny>=2.19.0",
  "psycopg>=3.2",
  "psycopg-pool>=3.2",
]

[project.license]
file = "LICENSE.txt"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
include = [
  "src/*"
]

[tool.hatch.build.targets.wheel]
packages = ["src/databricks_ai_bridge"]

[project.optional-dependencies]
dev = [
  "hatch",
  "pytest",
  "ruff==0.6.4",
  "databricks-vectorsearch>=0.50",
]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = [
  # isort
  "I",
  # bugbear rules
  "B",
  # remove unused imports
  "F401",
  # bare except statements
  "E722",
  # print statements
  "T201",
  "T203",
  # misuse of typing.TYPE_CHECKING
  "TCH004",
  # import rules
  "TID251",
  # undefined-local-with-import-star
  "F403",
]

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 88

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pytest.ini_options]
pythonpath = ["src"]
